sugested changes
